Moved prettyPrintJson() and examine() into debug.dart so they can be used anywhere.
